#1261 make multiSqlJob thread-safe & change meta reload count-down
