+++ src/main/java/com/actiontech/dble/sqlengine/MultiSQLJob.java
@@ -17,6 +17,7 @@ import com.actiontech.dble.route.RouteResultsetNode;
 import com.actiontech.dble.server.parser.ServerParse;
 
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * async execute in EngineCtx or standalone (EngineCtx=null)
@@ -34,7 +35,7 @@ public class MultiSQLJob implements ResponseHandler, Runnable {
     private final SQLJobHandler jobHandler;
     private final PhysicalDatasource ds;
     private boolean isMustWriteNode;
-    private volatile boolean finished;
+    private AtomicBoolean finished = new AtomicBoolean(false);
 
     public MultiSQLJob(String sql, String schema, SQLJobHandler jobHandler, PhysicalDatasource ds, boolean isReload) {
         super();
@@ -95,13 +96,14 @@ public class MultiSQLJob implements ResponseHandler, Runnable {
     }
 
     public boolean isFinished() {
-        return finished;
+        return finished.get();
     }
 
     private void doFinished(boolean failed) {
         logger.info("Finish MultiSQLJob with result " + failed + " on connection " + connection);
-        finished = true;
-        jobHandler.finished(dataNode == null ? schema : dataNode, failed);
+        if (finished.compareAndSet(false, true)) {
+            jobHandler.finished(dataNode == null ? schema : dataNode, failed);
+        }
     }
 
     @Override
